feat: make reads clean
